[FE] refactor: 여행 수정 모달 기능 개선
